+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-## Initial release
+## In this update
+- New `ignore` feature to skip certain chars